Fix clang warning
git-svn-id: https://svn.osgeo.org/gdal/trunk@37660 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/pds/isis3dataset.cpp
@@ -278,7 +278,7 @@ class CPLJsonObject
             m_oList.insert( m_oList.begin() + nPos,
                             std::pair<CPLString,CPLJsonObject*>(
                                             osKey, new CPLJsonObject(obj)) );
-            m_oMap[osKey] = nPos;
+            m_oMap[osKey] = static_cast<int>(nPos);
         }
 
         bool has(const CPLString& osKey) const
